+++ docs/Changelog.md
@@ -2,6 +2,7 @@
 
 ## 1.0.0-beta.3 (TBA, 2025)
 * Implemented End-to-End Encryption
+* Token Based Authentication.
 * Improved stability and internals.
 
 ## 1.0.0-beta.2 (May 18, 2025)